requirements.txt
asgiref==3.7.2
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
